Changed SQL to match TYPO3 requirements for double2 fields Added fake employe to holidays file for testing delete check Changed extension type to "example"
git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/externalimport_tut/trunk@14579 735d13b6-9817-0410-8766-e36946ffe9aa

+++ ext_tables.sql
@@ -24,5 +24,5 @@ CREATE TABLE tx_externalimporttut_departments (
 CREATE TABLE fe_users (
 	tx_externalimporttut_code varchar(10) DEFAULT '' NOT NULL,
 	tx_externalimporttut_department text,
-	tx_externalimporttut_holidays double DEFAULT '0.0' NOT NULL,
+	tx_externalimporttut_holidays double(11,2) DEFAULT '0.0' NOT NULL,
 );
